fix(sp1): modify lift logic in order to support single record
Previously, we followed the pattern in:
https://github.com/zk-rabbit/sp1-cluster/blob/bf1c2b2/crates/worker/src/tasks/prove_shard.rs#L260-L267

However, for single-record requests we could not cover the case properly,
since recursionWitnessValue never received is_complete = true (it bypassed compression).

As a result, verification failed with the error:
InvalidPublicValues("is_complete is not 1").

This commit fixes the issue by replacing the forced value with the proper logic.